unconvert linter

+++ router/throttler/ratelimiter/ratelimiter.go
@@ -71,7 +71,7 @@ func (r *RateLimiter) Check(key string, currentTime time.Time) (limitStatus *Lim
 	}
 	timeFromCurrWindow := currentTime.UTC().Sub(currentWindow)
 
-	rate := float64((float64(r.windowSize)-float64(timeFromCurrWindow))/float64(r.windowSize))*float64(prevValue) + float64(currentValue)
+	rate := (float64(r.windowSize)-float64(timeFromCurrWindow))/float64(r.windowSize)*float64(prevValue) + float64(currentValue)
 	limitStatus = &LimitStatus{}
 	if rate >= float64(r.requestsLimit) {
 		limitStatus.IsLimited = true
